third_party/WebKit/LayoutTests/web-animations-api/timeline-updates-players.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<!-- This tests that players get updated even if they have previously had their source content removed or go out of effect. -->

<div id='element'></div>

<script>
var element = document.getElementById('element');
var animation = new Animation(element, [{left: '0px'}, {left: '100px'}], 10);
test(function() {
    var player = document.timeline.play(animation);
    player.finish();
    assert_equals(getComputedStyle(element).left, 'auto');

    player.play();
    assert_equals(getComputedStyle(element).left, '0px');

    player.finish();
}, 'Player updated after finish and play');

test(function() {
    var player = document.timeline.play(animation);
    assert_equals(getComputedStyle(element).left, '0px');
    var source = player.source
    player.source = null;
    assert_equals(getComputedStyle(element).left, 'auto');

    player.source = source;
    assert_equals(getComputedStyle(element).left, '0px');

    player.finish();
}, 'Player updated after clear and set source');
</script>
